+++ src/hooks/useParamsDataSource.js
@@ -2,12 +2,12 @@ import { useQuery } from 'react-query';
 
 export const useParamsDataSource = ({ source, searchValue, getParamsSource }) => {
   const { data, isLoading } = useQuery({
-    queryKey: [source?.source?.entityTypeId, searchValue],
-    queryFn: () => getParamsSource(
-      { entityTypeId: source?.source.entityTypeId,
-        source: source?.source.columnName,
-        search: searchValue },
-    ),
+    queryKey: ['getParamsSource', source, searchValue],
+    queryFn: () => getParamsSource({
+      entityTypeId: source?.source.entityTypeId,
+      source: source?.source.columnName,
+      search: searchValue,
+    }),
     enabled: !!source,
   });