+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 * [UIPQB-17](https://issues.folio.org/browse/UIPQB-17) Test query
 * [UIPQB-28](https://issues.folio.org/browse/UIPQB-28) Update QueryBuilder Component to work with MongoDB-Compliant JSON Query
 * [UIPQB-27](https://issues.folio.org/browse/UIPQB-27) Title: Extend QueryBuilder plugin with 'saveBtnLabel' Prop
+* [UIPQB-30](https://issues.folio.org/browse/UIPQB-30) Create pulling mechanism for query-builder-plugin